pyproject.toml
[tool.poetry]
name = "pdfshell"
version = "0.1.0"
description = "Plugin-driven, language-aware PDF automation platform"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "core" },
    { include = "cli" },
    { include = "tools" },
    { include = "agent" },
    { include = "coreapi" },
    { include = "trace" },
    { include = "pdfshell_srv" },
    { include = "pdfshell/shell" }
]

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
pypdf = "^4.2.0"
PyYAML = "^6.0.1"
requests = "^2.31.0"
django = "^5.0.6"
djangorestframework = "^3.15.1"
pydantic = "2.11.3"
gunicorn = "^21.2.0"
langchain = "0.3.24"
langchain-core = "0.3.55"
langchain-openai = "^0.2.0"
openai = ">=1.26.0,<2.0.0"
reportlab = "^4.1.0"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.0.0"
Pillow = "^10.0.0"
python-dotenv = "^1.0.1"
click-shell = "^2.1"
rich = "^13"
langgraph = "^0.4.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.0.0"
black = "^24.0.0"
flake8 = "^7.0.0"
Pillow = "^10.0.0"
pytest-django = "^4.8.0"

[tool.poetry.extras]
api = ["django", "djangorestframework", "pydantic", "gunicorn", "PyYAML", "psycopg2-binary", "dj-database-url"]
pdf = ["reportlab", "Pillow"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pdfshell_cli = "cli.main:merge"
pdfshell_interactive = "pdfshell.shell.app:pdfshell"

